+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Keep it human-readable, your future self will thank you!
 
 - Call filters from anemoi-transform
 - make test optional when adls is not installed Pull request #110
+- Add wz_to_w, orog_to_z, and sum filters (#149)
 
 ## [0.5.8](https://github.com/ecmwf/anemoi-datasets/compare/0.5.7...0.5.8) - 2024-10-26